Extract code from ProgressSection to ActionsArea
Extract the parts of the code for the ProgressSection component that relate to actions, and put them into a new component called ActionsArea. Also move the `latestActions` state into the game context.
